refactor : getToday 함수 utils로 분류
